feat: epic-grouped features tab + consolidated resources tab
- Restructure features-tab to nest features under their parent epic (accordion)
- Each epic is expandable with progress indicator and colored border
- Standalone features shown in separate section
- Merge Documents + Links into a single Resources tab
- Delete documents-tab.tsx and links-tab.tsx
- Update project-detail.tsx to use new tab structure
